requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
lxml
redis
prometheus-fastapi-instrumentator
prometheus-client
tenacity
python-dotenv
pandas
elasticsearch
python-multipart
slowapi
pytest
pytest-asyncio
pytest-cov
pydantic-settings
celery[redis]>=5.3.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
aiosqlite>=0.19.0
psutil
psycopg2-binary
email-validator
jinja2

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Performance
bitarray>=2.0.0

# Monitoring
sentry-sdk[fastapi]>=2.0.0

# Admin Panel
fastapi-admin>=2.0.0
sqladmin>=0.19.0

